refactor: remove user model
