+++ backend/src/modules/auth/interfaces/services/validate-user.auth.service.interface.ts
@@ -1,9 +1,9 @@
-import UserModel from 'src/modules/users/entities/user';
+import User from 'src/modules/users/entities/user.schema';
 
 export interface ValidateUserAuthService {
-	validateUserWithCredentials(email: string, plainTextPassword: string): Promise<UserModel | null>;
+	validateUserWithCredentials(email: string, plainTextPassword: string): Promise<User | null>;
 
-	validateUserById(userId: string): Promise<UserModel | null>;
+	validateUserById(userId: string): Promise<User | null>;
 
-	validateUserByRefreshToken(authorization: string, userId: string): Promise<false | UserModel>;
+	validateUserByRefreshToken(authorization: string, userId: string): Promise<false | User>;
 }
