+++ backend/src/modules/auth/shared/login.auth.ts
@@ -1,10 +1,10 @@
 import UserDto from 'src/modules/users/dto/user.dto';
-import UserModel from 'src/modules/users/entities/user';
+import User from 'src/modules/users/entities/user.schema';
 import { GetTokenAuthApplication } from '../interfaces/applications/get-token.auth.application.interface';
 import { GetTokenAuthService } from '../interfaces/services/get-token.auth.service.interface';
 
 export const signIn = async (
-	user: UserModel | UserDto,
+	user: User | UserDto,
 	getTokenService: GetTokenAuthService | GetTokenAuthApplication,
 	strategy: string
 ) => {
