--- backend/src/modules/users/entities/user.ts
@@ -1,21 +1,0 @@
-export default class UserModel {
-	_id?: string;
-
-	firstName: string;
-
-	lastName: string;
-
-	password: string;
-
-	email: string;
-
-	strategy: string;
-
-	currentHashedRefreshToken?: string;
-
-	isSAdmin: boolean;
-
-	isDeleted: boolean;
-
-	joinedAt: Date;
-}
