+++ backend/src/modules/users/interfaces/services/create.user.service.interface.ts
@@ -1,6 +1,6 @@
 import CreateUserAzureDto from '../../dto/create.user.azure.dto';
 import CreateUserDto from '../../dto/create.user.dto';
-import User from '../../entities/user';
+import User from '../../entities/user.schema';
 
 export interface CreateUserService {
 	create(user: CreateUserDto | CreateUserAzureDto): Promise<User>;
