+++ backend/src/modules/users/interfaces/type-user-with-teams.ts
@@ -1,4 +1,4 @@
-import User from '../entities/user';
+import User from '../entities/user.schema';
 
 export type UserWithTeams = {
 	user: User;
