+++ backend/src/modules/users/repository/user.repository.interface.ts
@@ -1,4 +1,4 @@
 import { BaseInterfaceRepository } from 'src/libs/repositories/interfaces/base.repository.interface';
-import User from '../entities/user';
+import User from '../entities/user.schema';
 
 export type UserRepositoryInterface = BaseInterfaceRepository<User>;
