+++ backend/src/modules/users/repository/user.repository.ts
@@ -2,16 +2,22 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { MongoGenericRepository } from 'src/libs/repositories/mongo/mongo-generic.repository';
-import UserModel from '../entities/user';
 import User, { UserDocument } from '../entities/user.schema';
 import { UserRepositoryInterface } from './user.repository.interface';
 
 @Injectable()
 export class UserRepository
-	extends MongoGenericRepository<UserModel>
+	extends MongoGenericRepository<User>
 	implements UserRepositoryInterface
 {
 	constructor(@InjectModel(User.name) private model: Model<UserDocument>) {
 		super(model);
+	}
+
+	getById(userId: string) {
+		return this.get(userId, {
+			password: 0,
+			currentHashedRefreshToken: 0
+		});
 	}
 }
